split tests into two separated files
